movemos pasta teste e refizemos os testes
