+++ src/business/error/NotFoundError.ts
@@ -1,4 +1,4 @@
-import { BaseError } from './baseError';
+import { BaseError } from './BaseError';
 
 export class NotFoundError extends BaseError {
   constructor(message: string) {
